Adding slide to emphasis 4 key rules
Was talking to Brad last night about how Souders presentation
overwhelmed someone last year. I think the section does a good job of
showing what matters, but I wanted to add a slide the makes it clear
that while there are a lot of performance rules, the biggest impact
comes from the stuff we covered.